Move open-browser from utils
