+++ src/open-browser.js
